requirements.txt
Django==3.1
djangorestframework==3.11.1
factory-boy==3.0.1
pre-commit==2.7.1
psycopg2==2.8.5
pycpfcnpj==1.5.1
